+++ webview/htr/myscript.js
@@ -1,6 +1,11 @@
 function myscript(token) {
+    const mimeType = 'application/x-latex';
 
-    var getStrokeGroups = () => {
+    // create an empty model
+    const model = iink.InkModel.createModel();
+
+    // fill with the stroke groups
+    model.strokeGroups = (() => {
         let strokes = [];
         for (const item of svg.children) {
             if (item.nodeName === 'path') {
@@ -18,63 +23,32 @@ function myscript(token) {
             }
         }
         return [{ "penStyle": null, "strokes": strokes }];
-    };
+    })();
 
-    const iinkRecognizer = iink.DefaultBehaviors.recognizerList.find(x => {
-        const infos = x.getInfo();
-        return infos.protocol === 'REST';
-    });
-
-    // Create a empty model
-    const model = iink.InkModel.createModel();
-    // Filling the model with the stroke groups
-    model.strokeGroups = getStrokeGroups();
-
-    // Creating a recognizer context with the configuration attached
+    // create a recognizer context with the configuration attached
     const recognizerContext = iink.RecognizerContext.createEmptyRecognizerContext({
-        configuration: iink.DefaultConfiguration
-    });
-
-    recognizerContext.editor.configuration.recognitionParams = {
-        type: 'MATH',
-        protocol: 'REST',
-        server: {
-            scheme: 'https',
-            // host: 'webdemoapi.myscript.com',
-            host: 'cloud.myscript.com',
-            applicationKey: token.app,
-            hmacKey: token.hmac
-        },
-        iink: {
-            math: {
-                mimeTypes: [
-                    'application/x-latex',
-                ],
+        configuration: {
+            recognitionParams: {
+                type: 'MATH',
+                protocol: 'REST',
+                server: {
+                    scheme: 'https',
+                    host: 'cloud.myscript.com',
+                    applicationKey: token.app,
+                    hmacKey: token.hmac
+                },
+                iink: {
+                    math: {
+                        mimeTypes: [mimeType],
+                    }
+                }
             }
         }
-    };
+    });
 
-    // Assigning a theme to the document
-    recognizerContext.editor.theme = iink.DefaultTheme;
-
-    // Defining the behaviour on recognition result
-    const recognitionCallback = (err, x) => {
-        if (!err) {
-            Object.entries(x.exports).forEach(([mimeType, exportValue]) => {
-                drawAPI.unstable.editCurrentLine({
-                    control: 0,
-                    text: '$$' + x.exports[mimeType].trim() + '$$'
-                });
-            });
-        }
-    };
-
-    // Triggering the recognition
-    iinkRecognizer.export_(recognizerContext, model)
-        .then((values) => {
-            values.forEach((value) => {
-                recognitionCallback(undefined, value);
-            });
-        })
-        .catch(err => recognitionCallback(err, undefined));
+    // trigger the recognition
+    return iink.DefaultBehaviors.recognizerList
+        .find(x => x.getInfo().protocol === 'REST')
+        .export_(recognizerContext, model)
+        .then(function (r) { return r[0].exports[mimeType].trim(); });
 }